+++ src/engine/data_storage/visitors/date_visitor.h
@@ -10,21 +10,21 @@
 #include "engine/data_storage/visitors/visitor.h"
 
 struct StrMaxVisitor : public ColumnVisitor {
-    void Visit(const Int8Column&,        const std::vector<size_t>&, size_t) override { throw std::runtime_error("StrMaxVisitor: non-str column"); }
-    void Visit(const Int16Column&,       const std::vector<size_t>&, size_t) override { throw std::runtime_error("StrMaxVisitor: non-str column"); }
-    void Visit(const Int32Column&,       const std::vector<size_t>&, size_t) override { throw std::runtime_error("StrMaxVisitor: non-str column"); }
-    void Visit(const Int64Column&,       const std::vector<size_t>&, size_t) override { throw std::runtime_error("StrMaxVisitor: non-str column"); }
-    void Visit(const Int128Column&,      const std::vector<size_t>&, size_t) override { throw std::runtime_error("StrMaxVisitor: non-str column"); }
-    void Visit(const DoubleColumn&,      const std::vector<size_t>&, size_t) override { throw std::runtime_error("StrMaxVisitor: non-str column"); }
-    void Visit(const DateColumn&,        const std::vector<size_t>&, size_t) override { throw std::runtime_error("StrMaxVisitor: non-str column"); }
-    void Visit(const TimeStampColumn&,   const std::vector<size_t>&, size_t) override { throw std::runtime_error("StrMaxVisitor: non-str column"); }
+    void Visit(const Int8Column&,        const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("StrMaxVisitor: non-str column"); }
+    void Visit(const Int16Column&,       const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("StrMaxVisitor: non-str column"); }
+    void Visit(const Int32Column&,       const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("StrMaxVisitor: non-str column"); }
+    void Visit(const Int64Column&,       const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("StrMaxVisitor: non-str column"); }
+    void Visit(const Int128Column&,      const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("StrMaxVisitor: non-str column"); }
+    void Visit(const DoubleColumn&,      const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("StrMaxVisitor: non-str column"); }
+    void Visit(const DateColumn&,        const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("StrMaxVisitor: non-str column"); }
+    void Visit(const TimeStampColumn&,   const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("StrMaxVisitor: non-str column"); }
 
-    void Visit(const StrColumn& col, const std::vector<size_t>& group_indices, size_t max_group_index) override {
+    void Visit(const StrColumn& col, const std::vector<uint32_t>& group_indices, size_t max_group_index) override {
         const size_t n = std::min(col.Size(), group_indices.size());
         visitor_detail::EnsureCapacity(group_max_str, max_group_index);
         for (size_t j = 0; j < n; ++j) {
             size_t group_ind = group_indices[j];
-            if (group_ind == SIZE_MAX) continue;
+            if (group_ind == UINT32_MAX) continue;
             const auto elem = col.GetElemView(j);
             if (group_max_str[group_ind].empty() || elem > group_max_str[group_ind])
                 group_max_str[group_ind] = elem;
@@ -42,21 +42,21 @@ private:
 
 
 struct StrMinVisitor : public ColumnVisitor {
-    void Visit(const Int8Column&,        const std::vector<size_t>&, size_t) override { throw std::runtime_error("StrMinVisitor: non-str column"); }
-    void Visit(const Int16Column&,       const std::vector<size_t>&, size_t) override { throw std::runtime_error("StrMinVisitor: non-str column"); }
-    void Visit(const Int32Column&,       const std::vector<size_t>&, size_t) override { throw std::runtime_error("StrMinVisitor: non-str column"); }
-    void Visit(const Int64Column&,       const std::vector<size_t>&, size_t) override { throw std::runtime_error("StrMinVisitor: non-str column"); }
-    void Visit(const Int128Column&,      const std::vector<size_t>&, size_t) override { throw std::runtime_error("StrMinVisitor: non-str column"); }
-    void Visit(const DoubleColumn&,      const std::vector<size_t>&, size_t) override { throw std::runtime_error("StrMinVisitor: non-str column"); }
-    void Visit(const DateColumn&,        const std::vector<size_t>&, size_t) override { throw std::runtime_error("StrMinVisitor: non-str column"); }
-    void Visit(const TimeStampColumn&,   const std::vector<size_t>&, size_t) override { throw std::runtime_error("StrMinVisitor: non-str column"); }
+    void Visit(const Int8Column&,        const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("StrMinVisitor: non-str column"); }
+    void Visit(const Int16Column&,       const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("StrMinVisitor: non-str column"); }
+    void Visit(const Int32Column&,       const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("StrMinVisitor: non-str column"); }
+    void Visit(const Int64Column&,       const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("StrMinVisitor: non-str column"); }
+    void Visit(const Int128Column&,      const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("StrMinVisitor: non-str column"); }
+    void Visit(const DoubleColumn&,      const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("StrMinVisitor: non-str column"); }
+    void Visit(const DateColumn&,        const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("StrMinVisitor: non-str column"); }
+    void Visit(const TimeStampColumn&,   const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("StrMinVisitor: non-str column"); }
 
-    void Visit(const StrColumn& col, const std::vector<size_t>& group_indices, size_t max_group_index) override {
+    void Visit(const StrColumn& col, const std::vector<uint32_t>& group_indices, size_t max_group_index) override {
         const size_t n = std::min(col.Size(), group_indices.size());
         visitor_detail::EnsureCapacity(group_min_str, max_group_index);
         for (size_t j = 0; j < n; ++j) {
             size_t group_ind = group_indices[j];
-            if (group_ind == SIZE_MAX) continue;
+            if (group_ind == UINT32_MAX) continue;
             const auto elem = col.GetElemView(j);
             if (group_min_str[group_ind].empty() || elem < group_min_str[group_ind])
                 group_min_str[group_ind] = elem;
@@ -74,33 +74,33 @@ private:
 
 
 struct DateMaxVisitor : public ColumnVisitor {
-    void Visit(const Int8Column&,    const std::vector<size_t>&, size_t) override { throw std::runtime_error("DateMaxVisitor: non-date column"); }
-    void Visit(const Int16Column&,   const std::vector<size_t>&, size_t) override { throw std::runtime_error("DateMaxVisitor: non-date column"); }
-    void Visit(const Int32Column&,   const std::vector<size_t>&, size_t) override { throw std::runtime_error("DateMaxVisitor: non-date column"); }
-    void Visit(const Int64Column&,   const std::vector<size_t>&, size_t) override { throw std::runtime_error("DateMaxVisitor: non-date column"); }
-    void Visit(const Int128Column&,  const std::vector<size_t>&, size_t) override { throw std::runtime_error("DateMaxVisitor: non-date column"); }
-    void Visit(const DoubleColumn&,  const std::vector<size_t>&, size_t) override { throw std::runtime_error("DateMaxVisitor: non-date column"); }
-    void Visit(const StrColumn&,     const std::vector<size_t>&, size_t) override { throw std::runtime_error("DateMaxVisitor: non-date column"); }
+    void Visit(const Int8Column&,    const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("DateMaxVisitor: non-date column"); }
+    void Visit(const Int16Column&,   const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("DateMaxVisitor: non-date column"); }
+    void Visit(const Int32Column&,   const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("DateMaxVisitor: non-date column"); }
+    void Visit(const Int64Column&,   const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("DateMaxVisitor: non-date column"); }
+    void Visit(const Int128Column&,  const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("DateMaxVisitor: non-date column"); }
+    void Visit(const DoubleColumn&,  const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("DateMaxVisitor: non-date column"); }
+    void Visit(const StrColumn&,     const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("DateMaxVisitor: non-date column"); }
 
-    void Visit(const DateColumn& col, const std::vector<size_t>& group_indices, size_t max_group_index) override {
+    void Visit(const DateColumn& col, const std::vector<uint32_t>& group_indices, size_t max_group_index) override {
         const auto& data = col.Data();
         const size_t n = std::min(data.size(), group_indices.size());
         visitor_detail::EnsureCapacity(group_max_date, max_group_index, INT32_MIN);
         for (size_t j = 0; j < n; ++j) {
             size_t group_ind = group_indices[j];
-            if (group_ind == SIZE_MAX) continue;
+            if (group_ind == UINT32_MAX) continue;
             const int32_t x = DateToInt32(data[j]);
             if (x > group_max_date[group_ind]) group_max_date[group_ind] = x;
         }
     }
 
-    void Visit(const TimeStampColumn& col, const std::vector<size_t>& group_indices, size_t max_group_index) override {
+    void Visit(const TimeStampColumn& col, const std::vector<uint32_t>& group_indices, size_t max_group_index) override {
         const auto& data = col.Data();
         const size_t n = std::min(data.size(), group_indices.size());
         visitor_detail::EnsureCapacity(group_max_timestamp, max_group_index, INT64_MIN);
         for (size_t j = 0; j < n; ++j) {
             size_t group_ind = group_indices[j];
-            if (group_ind == SIZE_MAX) continue;
+            if (group_ind == UINT32_MAX) continue;
             const int64_t x = TimeStampToInt64(data[j]);
             if (x > group_max_timestamp[group_ind]) group_max_timestamp[group_ind] = x;
         }
@@ -122,33 +122,33 @@ private:
 
 
 struct DateMinVisitor : public ColumnVisitor {
-    void Visit(const Int8Column&,    const std::vector<size_t>&, size_t) override { throw std::runtime_error("DateMinVisitor: non-date column"); }
-    void Visit(const Int16Column&,   const std::vector<size_t>&, size_t) override { throw std::runtime_error("DateMinVisitor: non-date column"); }
-    void Visit(const Int32Column&,   const std::vector<size_t>&, size_t) override { throw std::runtime_error("DateMinVisitor: non-date column"); }
-    void Visit(const Int64Column&,   const std::vector<size_t>&, size_t) override { throw std::runtime_error("DateMinVisitor: non-date column"); }
-    void Visit(const Int128Column&,  const std::vector<size_t>&, size_t) override { throw std::runtime_error("DateMinVisitor: non-date column"); }
-    void Visit(const DoubleColumn&,  const std::vector<size_t>&, size_t) override { throw std::runtime_error("DateMinVisitor: non-date column"); }
-    void Visit(const StrColumn&,     const std::vector<size_t>&, size_t) override { throw std::runtime_error("DateMinVisitor: non-date column"); }
+    void Visit(const Int8Column&,    const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("DateMinVisitor: non-date column"); }
+    void Visit(const Int16Column&,   const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("DateMinVisitor: non-date column"); }
+    void Visit(const Int32Column&,   const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("DateMinVisitor: non-date column"); }
+    void Visit(const Int64Column&,   const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("DateMinVisitor: non-date column"); }
+    void Visit(const Int128Column&,  const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("DateMinVisitor: non-date column"); }
+    void Visit(const DoubleColumn&,  const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("DateMinVisitor: non-date column"); }
+    void Visit(const StrColumn&,     const std::vector<uint32_t>&, size_t) override { throw std::runtime_error("DateMinVisitor: non-date column"); }
 
-    void Visit(const DateColumn& col, const std::vector<size_t>& group_indices, size_t max_group_index) override {
+    void Visit(const DateColumn& col, const std::vector<uint32_t>& group_indices, size_t max_group_index) override {
         const auto& data = col.Data();
         const size_t n = std::min(data.size(), group_indices.size());
         visitor_detail::EnsureCapacity(group_min_date, max_group_index, INT32_MAX);
         for (size_t j = 0; j < n; ++j) {
             size_t group_ind = group_indices[j];
-            if (group_ind == SIZE_MAX) continue;
+            if (group_ind == UINT32_MAX) continue;
             const int32_t x = DateToInt32(data[j]);
             if (x < group_min_date[group_ind]) group_min_date[group_ind] = x;
         }
     }
 
-    void Visit(const TimeStampColumn& col, const std::vector<size_t>& group_indices, size_t max_group_index) override {
+    void Visit(const TimeStampColumn& col, const std::vector<uint32_t>& group_indices, size_t max_group_index) override {
         const auto& data = col.Data();
         const size_t n = std::min(data.size(), group_indices.size());
         visitor_detail::EnsureCapacity(group_min_timestamp, max_group_index, INT64_MAX);
         for (size_t j = 0; j < n; ++j) {
             size_t group_ind = group_indices[j];
-            if (group_ind == SIZE_MAX) continue;
+            if (group_ind == UINT32_MAX) continue;
             const int64_t x = TimeStampToInt64(data[j]);
             if (x < group_min_timestamp[group_ind]) group_min_timestamp[group_ind] = x;
         }
