+++ src/queries_executor/aggregation.h
@@ -15,7 +15,7 @@
 struct Aggregation {
     Aggregation(std::string column_name_, std::string result_name_ = "")
         : column_name(std::move(column_name_)), result_name(std::move(result_name_)) {}
-    virtual void RunBatch(const std::shared_ptr<Batch>&, const std::vector<size_t>& group_indices, size_t max_group_index) = 0;
+    virtual void RunBatch(const std::shared_ptr<Batch>&, const std::vector<uint32_t>& group_indices, size_t max_group_index) = 0;
     virtual std::string GetResultValue(size_t group_index) const = 0;
     virtual void GetResultInto(Column& column, size_t group_index) const = 0;
     virtual Type GetResultType() const = 0;
@@ -30,7 +30,7 @@ struct Aggregation {
 
 struct CountAggregation : public Aggregation {
     CountAggregation(std::string result_name_ = "");
-    void RunBatch(const std::shared_ptr<Batch>&, const std::vector<size_t>& group_indices, size_t max_group_index) override;
+    void RunBatch(const std::shared_ptr<Batch>&, const std::vector<uint32_t>& group_indices, size_t max_group_index) override;
     std::string GetResultValue(size_t group_index) const override;
     void GetResultInto(Column& column, size_t group_index) const override;
     Type GetResultType() const override;
@@ -42,7 +42,7 @@ struct CountAggregation : public Aggregation {
 
 struct SumAggregation : public Aggregation {
     SumAggregation(std::string col_name, std::string result_name = "");
-    void RunBatch(const std::shared_ptr<Batch>&, const std::vector<size_t>& group_indices, size_t max_group_index) override;
+    void RunBatch(const std::shared_ptr<Batch>&, const std::vector<uint32_t>& group_indices, size_t max_group_index) override;
     std::string GetResultValue(size_t group_index) const override;
     void GetResultInto(Column& column, size_t group_index) const override;
     Type GetResultType() const override;
@@ -54,7 +54,7 @@ struct SumAggregation : public Aggregation {
 
 struct AvgAggregation : public Aggregation {
     AvgAggregation(std::string col_name, std::string result_name = "");
-    void RunBatch(const std::shared_ptr<Batch>&, const std::vector<size_t>& group_indices, size_t max_group_index) override;
+    void RunBatch(const std::shared_ptr<Batch>&, const std::vector<uint32_t>& group_indices, size_t max_group_index) override;
     std::string GetResultValue(size_t group_index) const override;
     void GetResultInto(Column& column, size_t group_index) const override;
     Type GetResultType() const override;
@@ -66,7 +66,7 @@ struct AvgAggregation : public Aggregation {
 
 struct CountDistinctAggregation : public Aggregation {
     CountDistinctAggregation(std::string col_name, std::string result_name = "");
-    void RunBatch(const std::shared_ptr<Batch>&, const std::vector<size_t>& group_indices, size_t max_group_index) override;
+    void RunBatch(const std::shared_ptr<Batch>&, const std::vector<uint32_t>& group_indices, size_t max_group_index) override;
     std::string GetResultValue(size_t group_index) const override;
     void GetResultInto(Column& column, size_t group_index) const override;
     Type GetResultType() const override;
@@ -78,7 +78,7 @@ struct CountDistinctAggregation : public Aggregation {
 
 struct MaxAggregation : public Aggregation {
     MaxAggregation(std::string col_name, std::string result_name = "");
-    void RunBatch(const std::shared_ptr<Batch>&, const std::vector<size_t>& group_indices, size_t max_group_index) override;
+    void RunBatch(const std::shared_ptr<Batch>&, const std::vector<uint32_t>& group_indices, size_t max_group_index) override;
     std::string GetResultValue(size_t group_index) const override;
     void GetResultInto(Column& column, size_t group_index) const override;
     Type GetResultType() const override;
@@ -92,7 +92,7 @@ struct MaxAggregation : public Aggregation {
 
 struct MinAggregation : public Aggregation {
     MinAggregation(std::string col_name, std::string result_name = "");
-    void RunBatch(const std::shared_ptr<Batch>&, const std::vector<size_t>& group_indices, size_t max_group_index) override;
+    void RunBatch(const std::shared_ptr<Batch>&, const std::vector<uint32_t>& group_indices, size_t max_group_index) override;
     std::string GetResultValue(size_t group_index) const override;
     void GetResultInto(Column& column, size_t group_index) const override;
     Type GetResultType() const override;
